+++ angular.json
@@ -15,11 +15,11 @@
     "shared-styles": "libs/shared/styles",
     "ui": "apps/ui",
     "ui-core": "libs/ui/core",
+    "ui-database-graphql": "libs/ui/database/graphql",
     "ui-design-components-derivative": "libs/ui/design/components/derivative",
     "ui-design-components-elementary": "libs/ui/design/components/elementary",
     "ui-design-icons": "libs/ui/design/icons",
     "ui-e2e": "apps/ui-e2e",
-    "ui-features-trip": "libs/ui/features/trip",
-    "ui-graphql": "libs/ui/graphql"
+    "ui-features-trip": "libs/ui/features/trip"
   }
 }
