+++ apps/ui/src/app/imports/models/index.ts
@@ -1,3 +1,3 @@
 export { Person } from '@itinerary-assistant/domain';
 export { RouterPayload } from '@itinerary-assistant/ui/core';
-export { GetPersonQuery } from '@itinerary-assistant/ui/graphql';
+export { GetPersonQuery } from '@itinerary-assistant/ui/database/graphql';
