+++ apps/ui/src/app/imports/services/index.ts
@@ -1,3 +1,6 @@
 export { EndpointService, RouterService } from '@itinerary-assistant/ui/core';
 
-export { GetPersonGQL, GetPeopleGQL } from '@itinerary-assistant/ui/graphql';
+export {
+  GetPersonGQL,
+  GetPeopleGQL
+} from '@itinerary-assistant/ui/database/graphql';
