+++ libs/ui/core/src/lib/imports/modules/index.ts
@@ -1,1 +1,1 @@
-export { GraphQLModule } from '@itinerary-assistant/ui/graphql';
+export { GraphQLModule } from '@itinerary-assistant/ui/database/graphql';
