+++ libs/ui/database/graphql/README.md
@@ -1,0 +1,7 @@
+# ui-database-graphql
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test ui-database-graphql` to execute the unit tests.
