+++ libs/ui/database/graphql/project.json
@@ -1,14 +1,14 @@
 {
   "projectType": "library",
-  "root": "libs/ui/graphql",
-  "sourceRoot": "libs/ui/graphql/src",
+  "root": "libs/ui/database/graphql",
+  "sourceRoot": "libs/ui/database/graphql/src",
   "prefix": "itinerary-assistant",
   "targets": {
     "test": {
       "executor": "@nrwl/jest:jest",
-      "outputs": ["coverage/libs/ui/graphql"],
+      "outputs": ["coverage/libs/ui/database/graphql"],
       "options": {
-        "jestConfig": "libs/ui/graphql/jest.config.ts",
+        "jestConfig": "libs/ui/database/graphql/jest.config.ts",
         "passWithNoTests": true
       }
     },
@@ -16,8 +16,8 @@
       "executor": "@nrwl/linter:eslint",
       "options": {
         "lintFilePatterns": [
-          "libs/ui/graphql/**/*.ts",
-          "libs/ui/graphql/**/*.html"
+          "libs/ui/database/graphql/**/*.ts",
+          "libs/ui/database/graphql/**/*.html"
         ]
       }
     }
