+++ libs/ui/database/graphql/src/lib/generated-endpoints/index.ts
