+++ libs/ui/database/graphql/src/lib/operations/person-operations.gql
