+++ libs/ui/database/graphql/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../../../tsconfig.base.json",
+  "extends": "../../../../tsconfig.base.json",
   "files": [],
   "include": [],
   "references": [
