--- libs/ui/graphql/README.md
@@ -1,7 +1,0 @@
-# ui-graphql
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test ui-graphql` to execute the unit tests.
