+++ tsconfig.base.json
@@ -10,18 +10,12 @@
     "importHelpers": true,
     "target": "es2015",
     "module": "esnext",
-    "lib": [
-      "es2017",
-      "dom",
-      "esnext.asynciterable"
-    ],
+    "lib": ["es2017", "dom", "esnext.asynciterable"],
     "skipLibCheck": true,
     "skipDefaultLibCheck": true,
     "baseUrl": ".",
     "paths": {
-      "@itinerary-assistant/api/core": [
-        "libs/api/core/src/index.ts"
-      ],
+      "@itinerary-assistant/api/core": ["libs/api/core/src/index.ts"],
       "@itinerary-assistant/api/database/entities": [
         "libs/api/database/entities/src/index.ts"
       ],
@@ -40,17 +34,12 @@
       "@itinerary-assistant/api/features/trip": [
         "libs/api/features/trip/src/index.ts"
       ],
-      "@itinerary-assistant/api/shared": [
-        "libs/api/shared/src/index.ts"
-      ],
-      "@itinerary-assistant/domain": [
-        "libs/domain/src/index.ts"
-      ],
-      "@itinerary-assistant/shared/data": [
-        "libs/shared/data/src/index.ts"
-      ],
-      "@itinerary-assistant/ui/core": [
-        "libs/ui/core/src/index.ts"
+      "@itinerary-assistant/api/shared": ["libs/api/shared/src/index.ts"],
+      "@itinerary-assistant/domain": ["libs/domain/src/index.ts"],
+      "@itinerary-assistant/shared/data": ["libs/shared/data/src/index.ts"],
+      "@itinerary-assistant/ui/core": ["libs/ui/core/src/index.ts"],
+      "@itinerary-assistant/ui/database/graphql": [
+        "libs/ui/database/graphql/src/index.ts"
       ],
       "@itinerary-assistant/ui/design/components/derivative": [
         "libs/ui/design/components/derivative/src/index.ts"
@@ -63,15 +52,9 @@
       ],
       "@itinerary-assistant/ui/features/trip": [
         "libs/ui/features/trip/src/index.ts"
-      ],
-      "@itinerary-assistant/ui/graphql": [
-        "libs/ui/graphql/src/index.ts"
       ]
     },
     "allowSyntheticDefaultImports": true
   },
-  "exclude": [
-    "node_modules",
-    "tmp"
-  ]
+  "exclude": ["node_modules", "tmp"]
 }
